# expression: implement vectorized evaluation for `builtinJSONExtractSig`